requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
yfinance
pytesseract
pillow
pdf2image
python-docx
googletrans==3.1.0a0
pdfplumber
opencv-python
pymongo
